Add `without_pricelist` scope to `Spree::Price`
A new scope `without_pricelist` in the `Spree::Price` model to allow filtering of prices that are not associated with any price list.
The purpose of this change is to allow the system to differentiate between prices that are managed by price lists and those that are not.
This scope is useful for filtering and querying prices that should be treated as standalone (i.e., not tied to a price list).