Add `Spree::Variant::PriceSelectorByUserGroup` to manage price selection
The class Spree::Variant::PriceSelectorByUserGroup, which selects the appropriate price for a product variant
based on user group pricing options. The primary reason for this change is to allow for user group-specific pricing
by incorporating a class that can dynamically select the correct price based on user group and other pricing options
(such as country and currency). This facilitates the implementation of price differentiation for various customer segments.

The class can be used by setting
```
Spree.config do |config|
 config.variant_price_selector_class = 'Spree::Variant::PriceSelectorByUserGroup
end
```
in `config/initializers/spree.rb`